Separate loans worker concerns
